+++ src/app.ts
@@ -1,5 +1,6 @@
 import { AutoLabelerSubscriber } from './subscribers/auto-labeler.js';
 import { Carson } from './carson.js';
+import { CommandsSubscriber } from './subscribers/commands.js';
 import { ConflictsNotifierSubscriber } from './subscribers/conflicts-notifier.js';
 import { IssueIntakeSubscriber } from './subscribers/issue-intake.js';
 import { LockOldIssuesSubscriber } from './subscribers/lock-old-issues.js';
@@ -16,6 +17,7 @@ import { WelcomeSubscriber } from './subscribers/welcome.js';
 
 export const carson = new Carson([
   new AutoLabelerSubscriber(),
+  new CommandsSubscriber(),
   new ConflictsNotifierSubscriber(),
   new IssueIntakeSubscriber(),
   new LockOldIssuesSubscriber(),